# apps/server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarGPT API",
        "version": "1.0.0",
        "description": "AI-powered car recommendation system using Ollama. Find your perfect car by describing your needs in natural language.",
        "contact": {
            "name": "CarGPT Support",
            "url": "https://github.com/lucapalomba/CarGPT"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "tags": [
        {
            "name": "Cars",
            "description": "Car recommendation and search operations"
        },

        {
            "name": "System",
            "description": "System health and session management"
        }
    ],
    "paths": {
        "/api/find-cars": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Find cars based on requirements",
                "description": "Analyzes user requirements and returns 3 personalized car recommendations using AI",
                "operationId": "findCars",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Preferred language for the response and market restriction (e.g., 'es-ES' for Spain)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "requirements"
                                ],
                                "properties": {
                                    "requirements": {
                                        "type": "string",
                                        "minLength": 10,
                                        "description": "Detailed description of car requirements",
                                        "example": "Looking for a spacious family car with large trunk for 3 suitcases, low fuel consumption, max budget €25,000"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response with car recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindCarsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid or missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "status": "fail",
                                    "message": "Describe your needs in more detail (at least 10 characters)"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - Ollama connection issue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "status": "error",
                                    "message": "Unable to connect to Ollama. Make sure it is running (ollama serve)"
                                }
                            }
                        }
                    }
                }
            }
        },


        "/api/get-conversations": {
            "get": {
                "tags": [
                    "Q&A"
                ],
                "summary": "Get all active conversations (Development Only)",
                "description": "Retrieves internal state of all active user conversations. Access is restricted to non-production environments.",
                "operationId": "getConversations",
                "responses": {
                    "200": {
                        "description": "List of active conversations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "conversations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },

        "/api/refine-search": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Refine existing suggestions",
                "description": "Updates the list of car suggestions based on user feedback. Pinned cars are included in the process and re-elaborated to ensure contextual consistency.",
                "operationId": "refineSearch",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Preferred language for the response and market restriction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "feedback"
                                ],
                                "properties": {
                                    "feedback": {
                                        "type": "string",
                                        "description": "User feedback to refine results",
                                        "example": "Too expensive, I want something under €20k"
                                    },
                                    "pinnedCars": {
                                        "type": "array",
                                        "items": {
                                        "$ref": "#/components/schemas/Car"
                                        },
                                        "description": "List of cars to keep in the results. These cars will be re-elaborated by the AI during the refinement process."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully refined suggestions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindCarsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reset-conversation": {
            "post": {
                "tags": [
                    "System"
                ],
                "summary": "Reset conversation",
                "description": "Clears the current session's conversation history",
                "operationId": "resetConversation",
                "responses": {
                    "200": {
                        "description": "Conversation reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Conversation reset"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "description": "Check if the server and Ollama are running correctly",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "Health status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Car": {
                "type": "object",
                "properties": {
                    "make": {
                        "type": "string",
                        "description": "Car manufacturer",
                        "example": "Skoda"
                    },
                    "model": {
                        "type": "string",
                        "description": "Car model",
                        "example": "Octavia Wagon"
                    },
                    "year": {
                        "type": "integer",
                        "description": "Production year",
                        "example": 2023
                    },
                    "body_type": {
                        "type": "string",
                        "description": "Body type (Hatchback, Sedan, SUV, Station Wagon, etc.)",
                        "example": "Station Wagon"
                    },
                    "configuration": {
                        "type": "string",
                        "description": "Trim level or engine designation",
                        "example": "2.0 TDI Style"
                    },
                    "precise_model": {
                        "type": "string",
                        "description": "Precise brand, model, year, and equipment level",
                        "example": "Skoda Octavia Wagon 2.0 TDI 2023 Style"
                    },
                    "price": {
                        "type": "string",
                        "description": "Price range when new",
                        "example": "25,000-30,000€"
                    },
                    "price_when_new": {
                        "type": "string",
                        "description": "Base price when new",
                        "example": "28,000€"
                    },
                    "type": {
                        "type": "string",
                        "description": "Car type/category",
                        "example": "Station Wagon"
                    },
                    "market_availability": {
                        "type": "string",
                        "description": "Availability in the user's market",
                        "example": "Available in the US market"
                    },
                    "vehicle_properties": {
                        "type": "object",
                        "description": "Dynamic properties of the car",
                        "additionalProperties": {
                            "type": "object",
                            "properties": {
                                "translatedLabel": {
                                    "type": "string"
                                },
                                "value": {
                                    "type": "string"
                                }
                            }
                        },
                        "example": {
                            "trunk_size": {
                                "translatedLabel": "Trunk Size",
                                "value": "640 liters"
                            },
                            "fuel_consumption": {
                                "translatedLabel": "Fuel Consumption",
                                "value": "5.2 l/100km"
                            }
                        }
                    },
                    "strengths": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of strengths",
                        "example": [
                            "Huge trunk",
                            "Reliable",
                            "Spacious"
                        ]
                    },
                    "weaknesses": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of weaknesses",
                        "example": [
                            "Conservative design",
                            "Average resale value"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "description": "Why this car is suitable",
                        "example": "Perfect for families thanks to the space and reliability"
                    },
                    "percentage": {
                        "type": "string",
                        "description": "Match percentage (0-100)",
                        "example": "95"
                    },
                    "constraints_satisfaction": {
                        "type": "object",
                        "description": "Satisfaction percentages for each constraint",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "example": {
                            "budget": "90, fits well within budget",
                            "trunk_space": "85, large trunk but slightly less than requested",
                            "fuel_efficiency": "80, average consumption"
                        }
                    },
                    "reasoning": {
                        "type": "string",
                        "description": "Detailed reasoning behind the selection",
                        "example": "This car matches your need for space and efficiency perfectly."
                    },
                    "pinned": {
                        "type": "boolean",
                        "description": "Whether this car is pinned by the user",
                        "example": false
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "url": { "type": "string" },
                                "thumbnail": { "type": "string" },
                                "title": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "FindCarsResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "conversationId": {
                        "type": "string",
                        "description": "Session ID for this conversation",
                        "example": "sess_abc123"
                    },
                    "analysis": {
                        "type": "string",
                        "description": "AI analysis of user requirements",
                        "example": "Based on your needs for a spacious family car..."
                    },
                    "user_market": {
                        "type": "string",
                        "description": "The market inferred from the user's language",
                        "example": "United States"
                    },
                    "cars": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Car"
                        },
                        "minItems": 3,
                        "maxItems": 3,
                        "description": "Array of 3 recommended cars"
                    },
                    "provider": {
                        "type": "string",
                        "description": "AI provider used",
                        "example": "ollama"
                    },
                    "searchIntent": {
                        "$ref": "#/components/schemas/SearchIntent"
                    },
                    "suggestions": {
                        "type": "object",
                        "description": "Raw car suggestions before elaboration",
                        "properties": {
                            "analysis": {
                                "type": "string"
                            },
                            "choices": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "pinned_cars": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "SearchIntent": {
                "type": "object",
                "description": "Parsed user search intent",
                "properties": {
                    "user_country": {
                        "type": "string",
                        "description": "ISO-3166-1 alpha-3 country code",
                        "example": "ITA"
                    },
                    "primary_focus": {
                        "type": "string",
                        "description": "Main focus of user request (3 adjectives max)",
                        "example": "spacious, reliable, economical"
                    },
                    "constraints": {
                        "type": "object",
                        "description": "Search constraints",
                        "additionalProperties": true,
                        "example": {
                            "budget": "20000 - 30000",
                            "must_have": ["large trunk", "low consumption"]
                        }
                    },
                    "interesting_properties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Properties to include in car details",
                        "example": ["trunk_volume", "fuel_consumption", "ncap_rating"]
                    }
                }
            },
            "ComparisonCategory": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Category name",
                        "example": "Performance"
                    },
                    "car1": {
                        "type": "string",
                        "description": "Car 1 evaluation",
                        "example": "Strong acceleration, 150hp engine (rating: 8/10)"
                    },
                    "car2": {
                        "type": "string",
                        "description": "Car 2 evaluation",
                        "example": "Moderate performance, 130hp engine (rating: 7/10)"
                    },
                    "winner": {
                        "type": "string",
                        "enum": [
                            "car1",
                            "car2",
                            "tie"
                        ],
                        "description": "Which car wins this category",
                        "example": "car1"
                    }
                }
            },
            "CompareResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "comparison": {
                        "type": "object",
                        "properties": {
                            "comparison": {
                                "type": "string",
                                "description": "Introduction to the comparison",
                                "example": "Comparing two excellent family cars..."
                            },
                            "categories": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ComparisonCategory"
                                },
                                "description": "Comparison across different categories"
                            },
                            "conclusion": {
                                "type": "string",
                                "description": "Final comparison conclusion",
                                "example": "Overall, Car 1 is better for performance while Car 2 excels in fuel economy"
                            }
                        }
                    }
                }
            },
            "AskResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "car": {
                        "type": "string",
                        "description": "Car that was asked about",
                        "example": "Skoda Octavia Wagon"
                    },
                    "question": {
                        "type": "string",
                        "description": "Question that was asked",
                        "example": "How much does annual maintenance cost?"
                    },
                    "answer": {
                        "type": "string",
                        "description": "AI-generated answer",
                        "example": "The annual maintenance cost for a Skoda Octavia Wagon typically ranges from €400-600..."
                    }
                }
            },
            "Alternative": {
                "type": "object",
                "properties": {
                    "make": {
                        "type": "string",
                        "example": "Volkswagen"
                    },
                    "model": {
                        "type": "string",
                        "example": "Golf Variant"
                    },
                    "reason": {
                        "type": "string",
                        "description": "Why this is a good alternative",
                        "example": "Same philosophy, slightly more compact"
                    },
                    "advantages": {
                        "type": "string",
                        "description": "Advantages over the original car",
                        "example": "Lower price, easier to park"
                    }
                }
            },
            "AlternativesResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "alternatives": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Alternative"
                        },
                        "minItems": 3,
                        "maxItems": 3,
                        "description": "Array of 3 alternative cars"
                    }
                }
            },
            "HealthResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "ok",
                            "degraded"
                        ],
                        "description": "Overall system status",
                        "example": "ok"
                    },
                    "ollama": {
                        "type": "string",
                        "enum": [
                            "connected",
                            "disconnected"
                        ],
                        "description": "Ollama connection status",
                        "example": "connected"
                    },
                    "model": {
                        "type": "string",
                        "description": "AI model being used",
                        "example": "ministral"
                    },
                    "active_conversations": {
                        "type": "integer",
                        "description": "Number of active conversations",
                        "example": 3
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "description": "Short error status ('fail' for 4xx, 'error' for 5xx)",
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "description": "Detailed error message",
                        "example": "Error description"
                    },
                    "stack": {
                        "type": "string",
                        "description": "Stack trace (only returned in development mode)"
                    }
                },
                "required": [
                    "success",
                    "status",
                    "message"
                ]
            }
        }
    }
}